# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "## TMF API Reference: TMF633 - Service Catalog Management\n\n### Release\
    \ : 18.5 - December 2018\n\nService Catalog API is one of Catalog Management API\
    \ Family. Service Catalog API goal is to provide a catalog of services. \n\n###\
    \ Resource \n- serviceCatalog\n\n### Operations\nService Catalog API performs\
    \ the following operations on the resource :\n- Retrieve an entity or a collection\
    \ of entities depending on filter criteria\n- Partial update of an entity (including\
    \ updating rules)\n- Create an entity (including default values and creation rules)\n\
    - Delete an entity (for administration purposes)\n- Manage notification of events"
  version: "4.0.0"
  title: "Service Catalog Management"
host: "serverRoot"
basePath: "/tmf-api/serviceCatalogManagement/v4/"
...
schemes:
- "https"
consumes:
- "application/json;charset=utf-8"
produces:
- "application/json;charset=utf-8"
paths:
  /serviceCatalog:
    get:
      tags:
      - "serviceCatalog"
      summary: "List or find ServiceCatalog objects"
      description: "This operation list or find ServiceCatalog entities"
      operationId: "listServiceCatalog"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to be provided in response"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "Requested index for start of resources to be provided in response"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Requested number of resources to be provided in response"
        required: false
        type: "integer"
      responses:
        200:
          description: "Success"
          headers:
            X-Total-Count:
              type: "integer"
              description: "Total number of items matching criteria"
            X-Result-Count:
              type: "integer"
              description: "Actual number of items returned in the response body"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceCatalog"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCatalog"
    post:
      tags:
      - "serviceCatalog"
      summary: "Creates a ServiceCatalog"
      description: "This operation creates a ServiceCatalog entity."
      operationId: "createServiceCatalog"
      parameters:
      - in: "body"
        name: "serviceCatalog"
        description: "The ServiceCatalog to be created"
        required: true
        schema:
          $ref: "#/definitions/ServiceCatalog_Create"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/ServiceCatalog"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCatalog"
  /serviceCatalog/{id}:
    get:
      tags:
      - "serviceCatalog"
      summary: "Retrieves a ServiceCatalog by ID"
      description: "This operation retrieves a ServiceCatalog entity. Attribute selection\
        \ is enabled for all first level attributes."
      operationId: "retrieveServiceCatalog"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceCatalog"
        required: true
        type: "string"
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to provide in response"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/ServiceCatalog"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCatalog"
    delete:
      tags:
      - "serviceCatalog"
      summary: "Deletes a ServiceCatalog"
      description: "This operation deletes a ServiceCatalog entity."
      operationId: "deleteServiceCatalog"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceCatalog"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCatalog"
    patch:
      tags:
      - "serviceCatalog"
      summary: "Updates partially a ServiceCatalog"
      description: "This operation updates partially a ServiceCatalog entity."
      operationId: "patchServiceCatalog"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceCatalog"
        required: true
        type: "string"
      - in: "body"
        name: "serviceCatalog"
        description: "The ServiceCatalog to be updated"
        required: true
        schema:
          $ref: "#/definitions/ServiceCatalog_Update"
      responses:
        200:
          description: "Updated"
          schema:
            $ref: "#/definitions/ServiceCatalog"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCatalog"
...
definitions:
...
  RelatedParty:
    type: "object"
    properties:
      id:
        type: "string"
        description: "Unique identifier of a related party"
      href:
        type: "string"
        description: "Reference of the related party, could be a party reference or\
          \ a party role reference"
      name:
        type: "string"
        description: "Name of the related party"
      role:
        type: "string"
        description: "Role of the related party."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
      '@referredType':
        type: "string"
        description: "The actual type of the target instance when needed for disambiguation."
    description: "RelatedParty reference. A related party defines party or party role\
      \ linked to a specific entity."
    example:
      '@referredType': "@referredType"
      role: "role"
      '@baseType': "@baseType"
      '@type': "@type"
      name: "name"
      id: "id"
      href: "href"
      '@schemaLocation': "@schemaLocation"
...
  ServiceCatalog:
    type: "object"
    properties:
      id:
        type: "string"
        description: "Unique identifier of the ServiceCatalog"
      href:
        type: "string"
        description: "Unique reference of the service catalog"
      description:
        type: "string"
        description: "Description of this catalog"
      lastUpdate:
        type: "string"
        format: "date-time"
        description: "Date and time of the last update"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status"
      name:
        type: "string"
        description: "Name of the service catalog"
      version:
        type: "string"
        description: "ServiceCatalog version"
      category:
        type: "array"
        description: "List of service categories associated with this catalog"
        items:
          $ref: "#/definitions/ServiceCategoryRef"
      relatedParty:
        type: "array"
        description: "List of parties or party roles related to this category"
        items:
          $ref: "#/definitions/RelatedParty"
      validFor:
        description: "The period for which the service catalog is valid"
        $ref: "#/definitions/TimePeriod"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    description: "The root entity for service catalog management.\nA service catalog\
      \ is a group of service specifications made available through service candidates\
      \ that an organization provides to the consumers (internal consumers like its\
      \ employees or B2B customers or B2C customers). \nA service catalog typically\
      \ includes name, description and time period that is valid for. It will have\
      \ a list of ServiceCandidate catalog items. A ServiceCandidate is an entity\
      \ that makes a ServiceSpecification available to a catalog.\nA ServiceCandidate\
      \ and its associated ServiceSpecification may be \"published\" - made visible\
      \ -in any number of ServiceCatalogs, or in none."
    example:
      lifecycleStatus: "lifecycleStatus"
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      '@type': "@type"
      description: "description"
      relatedParty:
      - '@referredType': "@referredType"
        role: "role"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "@schemaLocation"
      - '@referredType': "@referredType"
        role: "role"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "@schemaLocation"
      version: "version"
      '@baseType': "@baseType"
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      id: "id"
      href: "href"
      category:
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      '@schemaLocation': "http://example.com/aeiou"
  ServiceCatalog_Create:
    type: "object"
    required:
    - "name"
    properties:
      description:
        type: "string"
        description: "Description of this catalog"
      lastUpdate:
        type: "string"
        format: "date-time"
        description: "Date and time of the last update"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status"
      name:
        type: "string"
        description: "Name of the service catalog"
      version:
        type: "string"
        description: "ServiceCatalog version"
      category:
        type: "array"
        description: "List of service categories associated with this catalog"
        items:
          $ref: "#/definitions/ServiceCategoryRef"
      relatedParty:
        type: "array"
        description: "List of parties or party roles related to this category"
        items:
          $ref: "#/definitions/RelatedParty"
      validFor:
        description: "The period for which the service catalog is valid"
        $ref: "#/definitions/TimePeriod"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    description: "The root entity for service catalog management.\nA service catalog\
      \ is a group of service specifications made available through service candidates\
      \ that an organization provides to the consumers (internal consumers like its\
      \ employees or B2B customers or B2C customers). \nA service catalog typically\
      \ includes name, description and time period that is valid for. It will have\
      \ a list of ServiceCandidate catalog items. A ServiceCandidate is an entity\
      \ that makes a ServiceSpecification available to a catalog.\nA ServiceCandidate\
      \ and its associated ServiceSpecification may be \"published\" - made visible\
      \ -in any number of ServiceCatalogs, or in none.\nSkipped properties: id,href"
    example:
      lifecycleStatus: "lifecycleStatus"
      '@baseType': "@baseType"
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      '@type': "@type"
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      description: "description"
      category:
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      relatedParty:
      - '@referredType': "@referredType"
        role: "role"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "@schemaLocation"
      - '@referredType': "@referredType"
        role: "role"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "@schemaLocation"
      '@schemaLocation': "http://example.com/aeiou"
      version: "version"
  ServiceCatalog_Update:
    type: "object"
    properties:
      description:
        type: "string"
        description: "Description of this catalog"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status"
      name:
        type: "string"
        description: "Name of the service catalog"
      version:
        type: "string"
        description: "ServiceCatalog version"
      category:
        type: "array"
        description: "List of service categories associated with this catalog"
        items:
          $ref: "#/definitions/ServiceCategoryRef"
      relatedParty:
        type: "array"
        description: "List of parties or party roles related to this category"
        items:
          $ref: "#/definitions/RelatedParty"
      validFor:
        description: "The period for which the service catalog is valid"
        $ref: "#/definitions/TimePeriod"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    description: "The root entity for service catalog management.\nA service catalog\
      \ is a group of service specifications made available through service candidates\
      \ that an organization provides to the consumers (internal consumers like its\
      \ employees or B2B customers or B2C customers). \nA service catalog typically\
      \ includes name, description and time period that is valid for. It will have\
      \ a list of ServiceCandidate catalog items. A ServiceCandidate is an entity\
      \ that makes a ServiceSpecification available to a catalog.\nA ServiceCandidate\
      \ and its associated ServiceSpecification may be \"published\" - made visible\
      \ -in any number of ServiceCatalogs, or in none.\nSkipped properties: id,href,lastUpdate"
    example:
      lifecycleStatus: "lifecycleStatus"
      '@baseType': "@baseType"
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      '@type': "@type"
      name: "name"
      description: "description"
      category:
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      relatedParty:
      - '@referredType': "@referredType"
        role: "role"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "@schemaLocation"
      - '@referredType': "@referredType"
        role: "role"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "@schemaLocation"
      '@schemaLocation': "http://example.com/aeiou"
      version: "version"
...
  ServiceCategoryRef:
    type: "object"
    required:
    - "href"
    - "id"
    properties:
      id:
        type: "string"
        description: "Unique identifier of category"
      href:
        type: "string"
        description: "Hypertext Reference of the category"
      name:
        type: "string"
        description: "Name of the category."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
      '@referredType':
        type: "string"
        description: "The actual type of the target instance when needed for disambiguation."
    description: "The (service) category resource is used to group service candidates\
      \ in logical containers. Categories can contain other categories."
    example:
      '@referredType': "@referredType"
      '@baseType': "@baseType"
      '@type': "@type"
      name: "name"
      id: "id"
      href: "href"
      '@schemaLocation': "http://example.com/aeiou"
...
  TimePeriod:
    type: "object"
    properties:
      endDateTime:
        type: "string"
        format: "date-time"
        description: "End of the time period, using IETC-RFC-3339 format"
      startDateTime:
        type: "string"
        format: "date-time"
        description: "Start of the time period, using IETC-RFC-3339 format. If you\
          \ define a start, you must also define an end"
    description: "A period of time, either as a deadline (endDateTime only) a startDateTime\
      \ only, or both"
    example:
      startDateTime: "2000-01-23T04:56:07.000+00:00"
      endDateTime: "2000-01-23T04:56:07.000+00:00"
...
  Error:
    type: "object"
    required:
    - "code"
    - "reason"
    properties:
      code:
        type: "integer"
        description: "Application relevant detail, defined in the API or a common\
          \ list."
      reason:
        type: "string"
        description: "Explanation of the reason for the error which can be shown to\
          \ a client user."
      message:
        type: "string"
        description: "More details and corrective actions related to the error which\
          \ can be shown to a client user."
      status:
        type: "integer"
        description: "HTTP Error code extension"
      referenceError:
        type: "string"
        format: "uri"
        description: "URI of documentation describing the error."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class."
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name."
    description: "Used when an API throws an Error, typically with a HTTP error response-code\
      \ (3xx, 4xx, 5xx)"